requirements.txt
requests
python-dotenv
art
